Drop unnecessary extension of #klass and #association_class
