+++ lib/active_record/precount/reflection_extension.rb
@@ -1,4 +1,20 @@
 module ActiveRecord
+  module Associations
+    class CountLoader < SingularAssociation
+      # Not preloaded behaviour of count_loader association
+      # When this method is called, it will be N+1 query
+      def load_target
+        count_target = reflection.name.to_s.sub(/_count\z/, '').to_sym
+        @target = owner.association(count_target).count
+
+        loaded! unless loaded?
+        target
+      rescue ActiveRecord::RecordNotFound
+        reset
+      end
+    end
+  end
+
   module Reflection
     class CountLoaderReflection < AssociationReflection
       def initialize(name, scope, options, active_record)
@@ -6,6 +22,14 @@ module ActiveRecord
       end
 
       def macro; :count_loader; end
+
+      def association_class
+        ActiveRecord::Associations::CountLoader
+      end
+
+      def klass
+        @klass ||= active_record.send(:compute_type, options[:class_name] || name.to_s.sub(/_count\z/, '').singularize.classify)
+      end
     end
   end
